+++ src/ADN.Helpers/src/Data/IntegerHelper.cs
@@ -22,27 +22,5 @@ namespace ADN.Helpers.Data
             if (value > max) return max;
             return Convert.ToInt32(value);
         }
-
-        public static string UIntToString(uint value)
-        {
-            System.Text.StringBuilder output = new System.Text.StringBuilder();
-            while (value > 0)
-            {
-                output.Append((char)((value & 0xFF)));
-                value >>= 8;
-            }
-            return output.ToString();
-        }
-
-        public static uint StringToUInt(string value)
-        {
-            uint output = 0;
-
-            for (int i = 0; i < value.Length; i++)
-            {
-                output += ((uint)value[i] << (i * 8));
-            }
-            return output;
-        }
     }
 }